+++ src/Storage.ts
@@ -134,23 +134,24 @@ export default class Storage {
     }
 
     // Return true if the timeline and user data is stale and should be reloaded
-    static async isDataStale(dataDescriptor?: string | StorageKey): Promise<boolean> {
-        let seconds;
-        let logPrefix = `[isDataStale`;
-
-        if (dataDescriptor) {
-            logPrefix += ` ${dataDescriptor}`;
-            seconds = await this.secondsSinceLastUpdated(dataDescriptor as StorageKey);
-        } else {
-            seconds = await this.secondsSinceMostRecentToot();
-        }
-
-        logPrefix += `]`;
+    static async isDataStale(key?: StorageKey): Promise<boolean> {
         const numAppOpens = await this.getNumAppOpens();
         // const isTenthAppOpen = (await this.getNumAppOpens()) % this.getConfig().reloadFeaturesEveryNthOpen === 0;
+        let logPrefix = `[isDataStale`;
+        let seconds;
+
+        if (key) {
+            logPrefix += ` ${key}]`;
+            seconds = await this.secondsSinceLastUpdated(key);
+            if (seconds) console.debug(`${logPrefix} Found updatedAt that's ${seconds} seconds old`);
+        } else {
+            logPrefix += `]`;
+            seconds = await this.secondsSinceMostRecentToot();
+            if (seconds) console.log(`${logPrefix} Using secondsSinceMostRecentToot(): ${seconds} seconds old`);
+        }
 
         if (numAppOpens <= 1) {
-            console.debug(`${logPrefix} numAppOpens is ${JSON.stringify(numAppOpens)} (initial load; data not stale)`);
+            console.debug(`${logPrefix} numAppOpens is ${JSON.stringify(numAppOpens)} so initial load; data not stale (${seconds} seconds)`);
             return false;
         } if (!seconds) {
             console.debug(`${logPrefix} secondsSinceMostRecentToot() returned ${JSON.stringify(seconds)} (data is stale)`);
@@ -178,6 +179,7 @@ export default class Storage {
             console.debug(`[${key}] secondsSinceLastUpdated(): ${age}`);
             return age;
         } else {
+            console.debug(`[${key}] secondsSinceLastUpdated(): No stored object found at key '${key}'`);
             return null;
         }
     }